+++ index.js
@@ -8,6 +8,7 @@ module.exports = {
     'no-loops': require('./rules/no-loops'),
     'no-mutating-assign': require('./rules/no-mutating-assign'),
     'no-mutation': require('./rules/no-mutation'),
+    'no-nil': require('./rules/no-nil'),
     'no-this': require('./rules/no-this'),
     'no-throw': require('./rules/no-throw'),
     'no-unused-expression': require('./rules/no-unused-expression')
@@ -22,6 +23,7 @@ module.exports = {
         'fp/no-loops': 'error',
         'fp/no-mutating-assign': 'error',
         'fp/no-mutation': 'error',
+        'fp/no-nil': 'error',
         'fp/no-this': 'error',
         'fp/no-throw': 'error',
         'fp/no-unused-expression': 'error'